next commit

+++ src/flights/flight.controller.js
@@ -1,13 +1,16 @@
 
+import { promise } from 'zod';
 import { CityService } from '../city/city.service.js';
 import { envs } from '../config/enviroments/enviroments.js';
 import { httpClient } from '../config/plugins/http-client.plugin.js';
 import { AppError, catchAsync } from '../errors/index.js';
 import { validateFlight, validatePartialFlight } from './flight.schema.js';
 import { FlightService } from './flight.service.js';
+import { TicketService } from '../tickets/ticket.service.js';
 
 const flightService = new FlightService();
 const cityService = new CityService()
+const ticketService = new TicketService()
 
 export const findAllFlights = catchAsync(async (req, res, next) => {
   const flights = await flightService.findAll();
@@ -72,13 +75,22 @@ export const updateFlights = catchAsync(async (req, res, next) => {
 export const deleteFlights = catchAsync(async (req, res, next) => {
   const { id } = req.params;
 
-  //TODO: no se deberia poder eliminar un vuelo pendiente, si ese vuelo tiene ticketes vendidos
-
   const flight = await flightService.findOne(id, 'pending');
 
   if (!flight) {
     return next(new AppError(`can't find flight with id: ${id}`));
-  }  
+  } 
+
+  const ticket = await ticketService.findOneTicketByFlightId(id)
+
+  if(ticket){
+    return next(
+      new AppError(
+        `a flight cannot be deleted if tickets have been for it`,
+        400
+      )
+    )
+  }
 
   await flightService.delete(flight);
 
@@ -89,18 +101,20 @@ export const approveFlight = catchAsync(async(req, res, next) => {
   const { id } = req.params;
 
   const flight = await flightService.findOne(id, 'pending');
-
   if(!flight){
     return next(new AppError(`flight with id: ${id} not found!`, 404))
   }
 
-  const originCity = await cityService.findOneCity(flight.originId);
+  const originCityPromise = cityService.findOneCity(flight.originId);
+
+  const destinationCityPromise =  cityService.findOneCity(flight.destinationId)
+  
+  const [originCity, destinationCity] = await Promise.all([originCityPromise, destinationCityPromise])
+
 
   if (!originCity){
     return next(new AppError('city of origin does not exist'));
   }
-
-  const destinationCity = await cityService.findOneCity(flight.destinationId)
 
   if(!destinationCity){
     return next(new AppError("city of destiny doesn't exists"));
@@ -109,7 +123,6 @@ export const approveFlight = catchAsync(async(req, res, next) => {
   const weatherConditions = await httpClient.get(
       `https://api.openweathermap.org/data/2.5/weather?lat=${originCity.lat}&lon=${originCity.long}&appid=${envs.API_KEY_WEATHERMAP}`
   )
-
   if(weatherConditions.weather[0].main === 'Rain'){
     return next(
       new AppError('weather conditions do not meet the requeriments for tokeoff', 400)
